Implementation of Publish Subscribe Message Exchange interface
As a replacement for the current Iceoryx integration in Cyclone,
this introduces the Publish Subscribe Message Exchange (PSMX)
interface that allows implementing additional methods of pub-sub
data-exchange as an alternative to Cyclone's networked transport.
The PSMX interface allows to load implementations as plugins, so
there is no need for compile-time linking.

One of the PSMX implementations is the Iceoryx plugin, which allows
using Iceoryx as shared memory transport for Cyclone. For testing,
a Cyclone based PSMX implementation is included.

This commit also includes some rework of the loan mechanism, so
that it better integrates with the PSMX mechanism.

Co-authored-by: Dennis Potman <dennis@zettascale.tech>
Co-authored-by: Martijn Reicher <martijn.reicher@zettascale.tech>

Signed-off-by: Dennis Potman <dennis@zettascale.tech>